extract FadBiomassAttractor into an interface
